Validation Framework v1.0
This is an elegant jQuery plugin for client-side validation that works
with both standard html and ASP.NET controls. It can be as simple or
complex as desired. Simply pass the plugin the DOM element and criteria
to be validated against and it will present the user an elegant message
in the event of an error.